app/requirements.txt
asyncpg==0.27.0
sqlalchemy==1.4.47
aiohttp==3.8.4
bcrypt==4.0.1
python-dotenv==1.0.0
requests==2.31.0
pydantic==2.0.2
gunicorn==20.1.0
